api-db/requirements.txt
asgiref==3.8.1
asyncio==3.4.3
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
Werkzeug==3.0.2